src/risk_analysis/sasb_integration.py
from typing import List, Dict
from src.models import Risk

def integrate_sasb_materiality(risks: List[Risk], industry: str) -> Dict[str, List[Risk]]:
    # This is a simplified implementation. In a real-world scenario, you would use the actual SASB Materiality Map
    sasb_categories = {
        "Energy": [
            "GHG Emissions",
            "Air Quality",
            "Water Management",
            "Waste Management",
            "Ecological Impacts",
            "Employee Health & Safety",
            "Business Ethics & Transparency"
        ],
        # Add other industries and their material issues here
    }
    
    material_risks = {category: [] for category in sasb_categories.get(industry, [])}
    
    for risk in risks:
        for category in sasb_categories.get(industry, []):
            if category.lower() in risk.description.lower():
                material_risks[category].append(risk)
    
    return material_risks